+++ src/main/java/com/github/lwhite1/tablesaw/mapping/TimeMapUtils.java
@@ -1,8 +1,6 @@
 package com.github.lwhite1.tablesaw.mapping;
 
-import com.github.lwhite1.tablesaw.api.IntColumn;
-import com.github.lwhite1.tablesaw.api.LongColumn;
-import com.github.lwhite1.tablesaw.api.ShortColumn;
+import com.github.lwhite1.tablesaw.api.DoubleColumn;
 import com.github.lwhite1.tablesaw.api.TimeColumn;
 import com.github.lwhite1.tablesaw.columns.TimeColumnUtils;
 import com.github.lwhite1.tablesaw.columns.packeddata.PackedLocalTime;
@@ -12,30 +10,30 @@ import java.time.temporal.ChronoUnit;
 
 public interface TimeMapUtils extends TimeColumnUtils {
 
-    default LongColumn differenceInMilliseconds(TimeColumn column2) {
+    default DoubleColumn differenceInMilliseconds(TimeColumn column2) {
         return difference(column2, ChronoUnit.MILLIS);
     }
 
-    default LongColumn differenceInSeconds(TimeColumn column2) {
+    default DoubleColumn differenceInSeconds(TimeColumn column2) {
         return difference(column2, ChronoUnit.SECONDS);
     }
 
-    default LongColumn differenceInMinutes(TimeColumn column2) {
+    default DoubleColumn differenceInMinutes(TimeColumn column2) {
         return difference(column2, ChronoUnit.MINUTES);
     }
 
-    default LongColumn differenceInHours(TimeColumn column2) {
+    default DoubleColumn differenceInHours(TimeColumn column2) {
         return difference(column2, ChronoUnit.HOURS);
     }
 
-    default LongColumn difference(TimeColumn column2, ChronoUnit unit) {
-        LongColumn newColumn = LongColumn.create(name() + " - " + column2.name());
+    default DoubleColumn difference(TimeColumn column2, ChronoUnit unit) {
+        DoubleColumn newColumn = DoubleColumn.create(name() + " - " + column2.name());
 
         for (int r = 0; r < size(); r++) {
             int c1 = this.getInt(r);
             int c2 = column2.getInt(r);
             if (c1 == TimeColumn.MISSING_VALUE || c2 == TimeColumn.MISSING_VALUE) {
-                newColumn.append(IntColumn.MISSING_VALUE);
+                newColumn.append(DoubleColumn.MISSING_VALUE);
             } else {
                 newColumn.append(difference(c1, c2, unit));
             }
@@ -49,40 +47,40 @@ public interface TimeMapUtils extends TimeColumnUtils {
         return unit.between(value1, value2);
     }
 
-    default ShortColumn hour() {
-        ShortColumn newColumn = ShortColumn.create(name() + "[" + "hour" + "]");
+    default DoubleColumn hour() {
+        DoubleColumn newColumn = DoubleColumn.create(name() + "[" + "hour" + "]");
         for (int r = 0; r < size(); r++) {
             int c1 = getInt(r);
             if (c1 != TimeColumn.MISSING_VALUE) {
                 newColumn.append(PackedLocalTime.getHour(c1));
             } else {
-                newColumn.append(ShortColumn.MISSING_VALUE);
+                newColumn.append(DoubleColumn.MISSING_VALUE);
             }
         }
         return newColumn;
     }
 
-    default IntColumn minuteOfDay() {
-        IntColumn newColumn = IntColumn.create(name() + "[" + "minute-of-day" + "]");
+    default DoubleColumn minuteOfDay() {
+        DoubleColumn newColumn = DoubleColumn.create(name() + "[" + "minute-of-day" + "]");
         for (int r = 0; r < size(); r++) {
             int c1 = getInt(r);
             if (c1 != TimeColumn.MISSING_VALUE) {
                 newColumn.append(PackedLocalTime.getMinuteOfDay(c1));
             } else {
-                newColumn.append(IntColumn.MISSING_VALUE);
+                newColumn.append(DoubleColumn.MISSING_VALUE);
             }
         }
         return newColumn;
     }
 
-    default IntColumn secondOfDay() {
-        IntColumn newColumn = IntColumn.create(name() + "[" + "second-of-day" + "]");
+    default DoubleColumn secondOfDay() {
+        DoubleColumn newColumn = DoubleColumn.create(name() + "[" + "second-of-day" + "]");
         for (int r = 0; r < size(); r++) {
             int c1 = getInt(r);
             if (c1 != TimeColumn.MISSING_VALUE) {
                 newColumn.append(PackedLocalTime.getSecondOfDay(c1));
             } else {
-                newColumn.append(IntColumn.MISSING_VALUE);
+                newColumn.append(DoubleColumn.MISSING_VALUE);
             }
         }
         return newColumn;
