+++ src/test/java/com/github/lwhite1/tablesaw/api/LocalDateColumnTest.java
@@ -39,7 +39,7 @@ public class LocalDateColumnTest {
         column1.appendCell("12/24/1924");
         column1.appendCell("12-May-2015");
         column1.appendCell("14-Jan-2015");
-        ShortColumn c2 = column1.dayOfMonth();
+        DoubleColumn c2 = column1.dayOfMonth();
         assertEquals(23, c2.get(0));
         assertEquals(24, c2.get(1));
         assertEquals(12, c2.get(2));
@@ -52,7 +52,7 @@ public class LocalDateColumnTest {
         column1.appendCell("12/24/1924");
         column1.appendCell("12-May-2015");
         column1.appendCell("14-Jan-2015");
-        ShortColumn c2 = column1.monthValue();
+        DoubleColumn c2 = column1.monthValue();
         assertEquals(10, c2.get(0));
         assertEquals(12, c2.get(1));
         assertEquals(5, c2.get(2));
@@ -64,7 +64,7 @@ public class LocalDateColumnTest {
         column1.appendCell("2013-10-23");
         column1.appendCell("12/24/1924");
         column1.appendCell("12-May-2015");
-        ShortColumn c2 = column1.year();
+        DoubleColumn c2 = column1.year();
         assertEquals(2013, c2.get(0));
         assertEquals(1924, c2.get(1));
         assertEquals(2015, c2.get(2));
